api/requirements.txt
 # Database connectivity
psycopg2-binary>=2.9.9
Django==5.2.4
djangorestframework>=3.14.0

# Django additional packages (commonly needed)
django-cors-headers>=4.0.0
django-environ>=0.10.0


# Data processing - using versions with pre-compiled wheels
pandas>=2.1.0
numpy>=1.25.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0


# Utilities
python-dotenv==1.0.1
